Update problemPath in setUnitAndProblem - fixes bugs in publishing teacher supports and logging
