+++ src/index.tsx
@@ -7,8 +7,8 @@ import { appConfigSpec, appIcons, createStores } from "./app-config";
 import { AppConfigContext } from "./app-config-context";
 import { AppComponent } from "./components/app";
 import { AppConfigModel } from "./models/stores/app-config-model";
+import { setUnitAndProblem } from "./models/stores/stores";
 import { UserModel } from "./models/stores/user";
-import { setUnitAndProblem } from "./models/curriculum/unit";
 import { urlParams } from "./utilities/url-params";
 import { getAppMode } from "./lib/auth";
 import { Logger } from "./lib/logger";
